+++ packages/terra-button/tests/jest/Button.test.jsx
@@ -49,24 +49,6 @@ it('should render a button with type equal to submit', () => {
   expect(button).toMatchSnapshot();
 });
 
-it('should render an icon', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} isIconOnly text="isIconOnly" />);
-  expect(button).toMatchSnapshot();
-});
-
-it('should render an icon and a text', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} text="text" />);
-  expect(button).toMatchSnapshot();
-});
-
-it('should render text then an icon when reversed', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} text="text" isReversed />);
-  expect(button).toMatchSnapshot();
-});
-
 it('should render a Button with merged attributes', () => {
   /* eslint-disable react/forbid-component-props */
   const button = shallow(<Button className="TestClass" data-mock="data" text="text" style={{ height: '100px' }} />);
@@ -148,18 +130,6 @@ it('should set the span text', () => {
   expect(button.find('.text-only').text()).toEqual('text');
 });
 
-it('should render the icon as the first child', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} text="text" />);
-  expect(button.childAt(0).childAt(0).hasClass('icon-first')).toEqual(true);
-});
-
-it('should render in order of text / icon when reverse', () => {
-  const testElement = <img alt="icon" />;
-  const button = shallow(<Button icon={testElement} isReversed text="text" />);
-  expect(button.childAt(0).childAt(0).hasClass('text-first')).toEqual(true);
-});
-
 it('should render as an anchor tag when provided an href', () => {
   const button = shallow(<Button href="MockHref" text="text" />);
   expect(button.is('a')).toEqual(true);
@@ -183,16 +153,6 @@ it('should append styles passed in with attributes', () => {
   /* eslint-enable react/forbid-component-props */
 });
 
-it('should render an icon only button with text as a fallback title', () => {
-  const button = shallow(<Button isIconOnly text="Icon Only" />);
-  expect(button).toMatchSnapshot();
-});
-
-it('should render an icon only button with a custom title', () => {
-  const button = shallow(<Button isIconOnly text="Icon Only" title="Custom Title" />);
-  expect(button).toMatchSnapshot();
-});
-
 it('should render a utility button with text as a fallback title', () => {
   const button = shallow(<Button variant="utility" text="Icon Only" />);
   expect(button).toMatchSnapshot();
@@ -205,16 +165,6 @@ it('should render a utility button with a custom title', () => {
 
 it('should render a button with a title', () => {
   const button = shallow(<Button text="Button" title="Custom Title" />);
-  expect(button).toMatchSnapshot();
-});
-
-it('should render an icon only button with text as a fallback accessible label', () => {
-  const button = shallow(<Button isIconOnly text="Icon Only" />);
-  expect(button).toMatchSnapshot();
-});
-
-it('should render an icon only button with a custom accessible label', () => {
-  const button = shallow(<Button isIconOnly text="Icon Only" aria-label="Custom Accessible Label" />);
   expect(button).toMatchSnapshot();
 });