+++ packages/terra-core-docs/src/terra-dev-site/test/button/Variants/ActionButton.test.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import IconSquare from 'terra-icon/lib/icon/IconSquareSymbol';
 import classNames from 'classnames/bind';
-import Button from 'terra-button';
+import { IconButton } from 'terra-button';
 import styles from '../ButtonTestCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
 const ActionButton = () => (
   <div id="action" className={cx('button-wrapper')}>
-    <Button id="actionButton" text="Action" icon={<IconSquare />} variant="action" />
-    <Button id="actionButtonDisabled" text="Action Disabled" icon={<IconSquare />} variant="action" isDisabled />
+    <IconButton id="actionButton" text="Action" icon={<IconSquare a11yLabel="square" />} iconType="informative" variant="action" />
+    <IconButton id="actionButtonDisabled" text="Action Disabled" icon={<IconSquare a11yLabel="square" />} iconType="informative" variant="action" isDisabled />
   </div>
 );