+++ packages/terra-core-docs/src/terra-dev-site/test/toolbar/CenterAlignToolbar.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from 'terra-button';
+import { IconButton } from 'terra-button';
 import IconEdit from 'terra-icon/lib/icon/IconEdit';
 import IconAdd from 'terra-icon/lib/icon/IconAdd';
 import IconAttachment from 'terra-icon/lib/icon/IconAttachment';
@@ -7,8 +7,8 @@ import Toolbar from 'terra-toolbar';
 
 export default () => (
   <Toolbar align="center">
-    <Button text="Edit" variant="utility" icon={<IconEdit />} />
-    <Button text="Add" variant="utility" icon={<IconAdd />} />
-    <Button text="Attachment" variant="utility" icon={<IconAttachment />} />
+    <IconButton text="Edit" variant="utility" icon={<IconEdit a11yLabel="Edit" />} iconType="informative" />
+    <IconButton text="Add" variant="utility" icon={<IconAdd a11yLabel="Add" />} iconType="informative" />
+    <IconButton text="Attachment" variant="utility" icon={<IconAttachment a11yLabel="Attachment" />} iconType="informative" />
   </Toolbar>
 );