+++ packages/terra-action-header/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Replaced `terra-button` with `terra-icon-button`
 
 * Breaking Changes
   * Renamed `title` prop as `text` to avoid confusion with HTML attribute `title`.