+++ packages/terra-action-header/CHANGELOG.md
@@ -8,6 +8,11 @@
   * Renamed `title` prop as `text` to avoid confusion with HTML attribute `title`.
   * Removed default `h1` heading level.
 
+## 2.74.3 - (July 5, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.74.2 - (February 24, 2022)
 
 * Changed